Refactor custom token management components
- Replaced instances of ManageVaultChainCoin with ManageVaultCoin for consistency across CustomToken and CustomTokenResult components.
- Introduced ManageVaultCoin component to encapsulate coin management logic, improving code organization and reusability.
- Cleaned up unused imports and optimized component structure in ManageVaultChainCoinsPage.
