+++ core/ui/chain/coin/addCustomToken/CustomToken.tsx
@@ -1,7 +1,7 @@
 import { areEqualCoins, coinKeyToString } from '@core/chain/coin/Coin'
 
 import { useCoreViewState } from '../../../navigation/hooks/useCoreViewState'
-import { ManageVaultChainCoin } from '../../../vault/chain/manage/coin/ManageVaultChainCoin'
+import { ManageVaultCoin } from '../../../vault/chain/manage/coin/ManageVaultCoin'
 import { useCurrentVaultChainCoins } from '../../../vault/state/currentVaultCoins'
 import { CustomTokenResult } from './CustomTokenResult'
 
@@ -13,7 +13,7 @@ export const CustomToken = ({ address }: { address: string }) => {
   const coin = coins.find(coin => areEqualCoins(coin, { id: address, chain }))
 
   if (coin) {
-    return <ManageVaultChainCoin key={coinKeyToString(coin)} value={coin} />
+    return <ManageVaultCoin key={coinKeyToString(coin)} value={coin} />
   }
 
   return <CustomTokenResult address={address} />
