+++ core/ui/vault/chain/manage/coin/index.tsx
@@ -1,86 +1,23 @@
 import { chainTokens } from '@core/chain/coin/chainTokens'
-import { areEqualCoins, Coin, extractCoinKey } from '@core/chain/coin/Coin'
+import { areEqualCoins } from '@core/chain/coin/Coin'
 import { sortCoinsAlphabetically } from '@core/chain/coin/utils/sortCoinsAlphabetically'
-import { ChainEntityIcon } from '@core/ui/chain/coin/icon/ChainEntityIcon'
-import { getCoinLogoSrc } from '@core/ui/chain/coin/icon/utils/getCoinLogoSrc'
 import { useWhitelistedCoinsQuery } from '@core/ui/chain/coin/queries/useWhitelistedCoinsQuery'
-import {
-  useAddToCoinFinderIgnoreMutation,
-  useRemoveFromCoinFinderIgnoreMutation,
-} from '@core/ui/storage/coinFinderIgnore'
-import {
-  useCreateCoinMutation,
-  useDeleteCoinMutation,
-} from '@core/ui/storage/coins'
 import { useCurrentVaultChain } from '@core/ui/vault/chain/useCurrentVaultChain'
-import {
-  useCurrentVaultChainCoins,
-  useCurrentVaultCoins,
-} from '@core/ui/vault/state/currentVaultCoins'
-import { Switch } from '@lib/ui/inputs/switch'
+import { useCurrentVaultChainCoins } from '@core/ui/vault/state/currentVaultCoins'
 import { TextInput } from '@lib/ui/inputs/TextInput'
-import { HStack, VStack } from '@lib/ui/layout/Stack'
+import { VStack } from '@lib/ui/layout/Stack'
 import { List } from '@lib/ui/list'
-import { ListItem } from '@lib/ui/list/item'
 import { PageContent } from '@lib/ui/page/PageContent'
 import { PageHeader } from '@lib/ui/page/PageHeader'
 import { PageHeaderBackButton } from '@lib/ui/page/PageHeaderBackButton'
 import { PageHeaderTitle } from '@lib/ui/page/PageHeaderTitle'
 import { Text } from '@lib/ui/text'
 import { withoutDuplicates } from '@lib/utils/array/withoutDuplicates'
-import { FC, useMemo, useState } from 'react'
+import { useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { AddCustomTokenPrompt } from './AddCustomTokenPrompt'
-
-const CoinItem: FC<Coin> = coin => {
-  const currentVaultCoins = useCurrentVaultCoins()
-  const createCoin = useCreateCoinMutation()
-  const deleteCoin = useDeleteCoinMutation()
-  const addToCoinFinderIgnore = useAddToCoinFinderIgnoreMutation()
-  const removeFromCoinFinderIgnore = useRemoveFromCoinFinderIgnoreMutation()
-
-  const currentVaultCoin = useMemo(() => {
-    return currentVaultCoins.find(c => areEqualCoins(c, coin))
-  }, [currentVaultCoins, coin])
-
-  const handleChange = () => {
-    if (currentVaultCoin) {
-      addToCoinFinderIgnore.mutate(extractCoinKey(currentVaultCoin), {
-        onSuccess: () => deleteCoin.mutate(currentVaultCoin),
-      })
-    } else {
-      removeFromCoinFinderIgnore.mutate(extractCoinKey(coin), {
-        onSuccess: () => createCoin.mutate(coin),
-      })
-    }
-  }
-
-  return (
-    <ListItem
-      extra={
-        <Switch
-          checked={!!currentVaultCoin}
-          onChange={handleChange}
-          loading={createCoin.isPending || deleteCoin.isPending}
-        />
-      }
-      icon={
-        <ChainEntityIcon
-          value={coin.logo ? getCoinLogoSrc(coin.logo) : undefined}
-          style={{ fontSize: 32 }}
-        />
-      }
-      title={
-        <HStack gap={12} alignItems="center">
-          <Text color="contrast" size={14} weight={500}>
-            {coin.ticker}
-          </Text>
-        </HStack>
-      }
-    />
-  )
-}
+import { ManageVaultCoin } from './ManageVaultCoin'
 
 export const ManageVaultChainCoinsPage = () => {
   const { t } = useTranslation()
@@ -148,7 +85,7 @@ export const ManageVaultChainCoinsPage = () => {
             </Text>
             <List>
               {activeCoins.map((coin, index) => (
-                <CoinItem key={index} {...coin} />
+                <ManageVaultCoin key={index} value={coin} />
               ))}
             </List>
           </VStack>
@@ -160,7 +97,7 @@ export const ManageVaultChainCoinsPage = () => {
             </Text>
             <List>
               {availableCoins.map((coin, index) => (
-                <CoinItem key={index} {...coin} />
+                <ManageVaultCoin key={index} value={coin} />
               ))}
             </List>
           </VStack>
